Move GeoNames support into shared infrastructure
